Microservices/transaction-service/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
pymysql
